+++ Assets/JHC/Prefab/BulletEffect/BulletEffect.prefab
@@ -4986,7 +4986,7 @@ Rigidbody2D:
     m_Bits: 0
   m_Interpolate: 0
   m_SleepingMode: 1
-  m_CollisionDetection: 0
+  m_CollisionDetection: 1
   m_Constraints: 0
 --- !u!1 &1666296132266488
 GameObject:
