+++ Assets/Scenes/PlayScene/NavMesh-NavMesh.asset.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: d52d1bf2a041bf245b1bbc29dd56706c
-folderAsset: yes
-DefaultImporter:
+guid: 5e61a8113d4901d469e62352e8a6a99b
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 23800000
   userData: 
   assetBundleName: 
   assetBundleVariant:
