+++ Assets/YJK/Glass.prefab.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: b41ad161d1889b848a121496cc2ef2a6
-DefaultImporter:
+guid: 9ae02db07cfb25144a721bed0c18b846
+PrefabImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
