+++ Assets/YJK/PlayerShoot.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4e7b39c5a032cc043ab13849a16b3dec
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
