+++ Assets/YJK/YJKPlayScene03.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 854937725fd01844e8986bc2b738e838
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
