+++ Assets/YJK/YJKPlayScene03.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 184e65cc1c915244aa19f227d12e2e49
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
