+++ Assets/YJK/YJKPlayScene03/NavMesh-NavMesh.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a9522453d7470084c8003d5b75713386
+guid: b38b7eac15e961a4f9e2b7481173c0b0
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 23800000
